+++ Functions/FillRand.cpp
@@ -1,0 +1,78 @@
+#include"FillRand.h"
+
+void FillRand(int arr[], const int n)
+{
+	for (int i = 0; i < n; i++)
+		arr[i] = rand() % 100;
+}
+void FillRand(double arr[], const int n)
+{
+	for (int i = 0; i < n; i++)
+		arr[i] = double(rand() % 100) / 10;
+}
+void FillRand(float arr[], const int n)
+{
+	for (int i = 0; i < n; i++)
+		arr[i] = float(rand() % 100) / 10;
+}
+void FillRand(char arr[], const int n)
+{
+	for (int i = 0; i < n; i++)
+		arr[i] = char(rand());
+}
+void FillRand(short arr[], const int n)
+{
+	for (int i = 0; i < n; i++)
+		arr[i] = short(rand() % 100);
+}
+
+void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			arr[i][j] = rand() % 100;
+		}
+	}
+}
+void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			arr[i][j] = double(rand() % 100) / 10;
+		}
+	}
+}
+void FillRand(float arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			arr[i][j] = float(rand() % 100) / 10;
+		}
+	}
+}
+void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			arr[i][j] = (char)rand();
+		}
+	}
+}
+void FillRand(short arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			arr[i][j] = short(rand() % 100);
+		}
+	}
+}
