+++ Functions/FillRand.h
@@ -1,0 +1,14 @@
+#pragma once
+#include"stdafx.h"
+
+void FillRand(int arr[], const int n);
+void FillRand(double arr[], const int n);
+void FillRand(float arr[], const int n);
+void FillRand(char arr[], const int n);
+void FillRand(short arr[], const int n);
+
+void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS);
+void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS);
+void FillRand(float arr[ROWS][COLS], const int ROWS, const int COLS);
+void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS);
+void FillRand(short arr[ROWS][COLS], const int ROWS, const int COLS);
