+++ Functions/Print.h
@@ -1,0 +1,8 @@
+#pragma once
+#include"stdafx.h"
+
+template<typename T>void Print(T arr[], const int n);
+
+template<typename T>void Print(T arr[ROWS][COLS], const int ROWS, const int COLS);
+
+template<typename T>void PrintReverse(T arr[], const int n);
