+++ Functions/main.cpp
@@ -1,44 +1,9 @@
-﻿#include<iostream>
-using namespace std;
+﻿#include"stdafx.h"
+#include"FillRand.h"
+#include"Print.h"
+#include"Statistics.h"
+#include"Sort.h"
 
-#define tab "\t"
-
-int g_a; //глобальная переменная, ее видят все функции, ее может изменить любая функция
-const int ROWS = 5;
-const int COLS = 8;
-
-template<typename T>void FillRand(T arr[], const int n);
-void FillRand(double arr[], const int n);
-void FillRand(float arr[], const int n);
-void FillRand(char arr[], const int n);
-void FillRand(short arr[], const int n);
-
-void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS);
-void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS);
-void FillRand(float arr[ROWS][COLS], const int ROWS, const int COLS);
-void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS);
-void FillRand(short arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>void Print(T arr[], const int n);
-
-template<typename T>void Print(T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>void PrintReverse(T arr[], const int n);
-
-template<typename T>T Sum(T arr[], const int n);
-template<typename T>T Sum(T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>double Avg(T arr[], const int n);
-template<typename T>double Avg(T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>T minValuein(T arr[], const int n);
-template<typename T>T minValuein(T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>T maxValuein(T arr[], const int n);
-template<typename T>T maxValuein(T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>void Sort(T arr[], const int n);
-template<typename T>void Sort(T arr[ROWS][COLS], const int ROWS, const int COLS);
 
 void main()
 {
@@ -147,224 +112,5 @@ void main()
 	cout << "Среднее арифметическое двумерного массива: " << Avg(s_arr_2, ROWS, COLS) << endl;
 	cout << "Минимальное значение из двумерного массива: " << minValuein(s_arr_2, ROWS, COLS) << endl;
 	cout << "Максимальное значение из двумерного массива: " << maxValuein(s_arr_2, ROWS, COLS) << endl << endl;
-	
-}
 
-template<typename T>void FillRand(T arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-		arr[i] = rand() % 100;
-}
-
-void FillRand(double arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-		arr[i] = double(rand() % 100) / 10;
-}
-void FillRand(float arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-		arr[i] = float(rand() % 100) / 10;
-}
-void FillRand(char arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-		arr[i] = char(rand());
-}
-void FillRand(short arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-		arr[i] = short(rand() % 100);
-}
-
-void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = rand() % 100;
-		}
-	}
-}
-void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = double(rand() % 100) / 10;
-		}
-	}
-}
-void FillRand(float arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = float(rand() % 100) / 10;
-		}
-	}
-}
-void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = (char)rand();
-		}
-	}
-}
-void FillRand(short arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = short(rand() % 100);
-		}
-	}
-}
-
-template<typename T>void Print(T arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-	{
-		cout << arr[i] << tab;
-	}
-	cout << endl;
-}
-
-template<typename T>void Print(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			cout << arr[i][j] << tab;
-		}
-		cout << endl;
-	}
-}
-
-template<typename T>void PrintReverse(T arr[], const int n)
-{
-	for (int i = n - 1; i >= 0; i--)
-	{
-		cout << arr[i] << tab;
-	}
-	cout << endl;
-}
-
-template<typename T>T Sum(T arr[], const int n)
-{
-	T sum = 0;
-	for (int i = 0; i < n; i++)
-	{
-		sum += arr[i];
-	}
-	return sum;
-
-}
-template<typename T>T Sum(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T sum = 0;
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			sum += arr[i][j];
-		}
-	}
-	return sum;
-}
-
-template<typename T>double Avg(T arr[], const int n)
-{
-	return (double)Sum(arr, n) / n;
-}
-template<typename T>double Avg(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	return (double)Sum(arr, ROWS, COLS) / (ROWS * COLS);
-}
-
-template<typename T>T minValuein(T arr[], const int n)
-{
-	T min = arr[0];
-	for (int i = 0; i < n; i++)
-		if (arr[i] < min) min = arr[i];
-	return min;
-
-}
-template<typename T>T minValuein(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T min = arr[0][0];
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			 (min > arr[i][j]) ? min = arr[i][j]: min;
-		}
-	}
-	return min;
-}
-
-template<typename T>T maxValuein(T arr[], const int n)
-{
-	T max = arr[0];
-	for (int i = 0; i < n; i++)
-		if (arr[i] > max) max = arr[i];
-	return max;
-}
-template<typename T>T maxValuein(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T max = arr[0][0];
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			(max < arr[i][j]) ? max = arr[i][j] : max;
-		}
-	}
-	return max;
-}
-
-template<typename T>void Sort(T arr[], const int n)
-{
-	T min = arr[0];
-	for (int i = 0; i < n; i++)
-	{
-		for (int j = i; j < n; j++)
-		{
-			if (arr[j] < min)
-			{
-				arr[i] = arr[j];
-				arr[j] = min;
-				min = arr[i];
-			}
-		}
-		min = arr[i + 1];
-	}
-}
-template<typename T>void Sort(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T min = arr[0][0];
-	for (int x = 0; x < ROWS; x++)
-	{
-		for (int i = 0; i < COLS; i++)
-		{
-			for (int j = i; j < COLS; j++)
-			{
-				if (arr[x][j] < min)
-				{
-					arr[x][i] = arr[x][j];
-					arr[x][j] = min;
-					min = arr[x][i];
-				}
-			}
-			min = arr[x][i+1];
-		}
-	}
 }
