+++ Functions/stdafx.h
@@ -1,0 +1,9 @@
+#pragma once
+#include<iostream>
+using namespace std;
+
+#define tab "\t"
+
+extern int g_a; //глобальная переменная, ее видят все функции, ее может изменить любая функция
+const int ROWS = 5;
+const int COLS = 8;
